fix: WASM compilation errors and publish to npm
- Fixed GpuOps trait implementation for WebGpuBackend
- Fixed arithmetic overflow in distributed module
- Added conditional VectorBackend::new calls for wasm32
- Added get_next_idx method to WasmVectorBackend
- Fixed Result type handling for wasm32 search methods
- Published vecstore-wasm@1.0.0 to npm

All 21 WASM compilation errors resolved.

+++ src/distributed/mod.rs
@@ -403,7 +403,7 @@ impl DistributedStore {
             status: NodeStatus::Joining,
             last_heartbeat: current_timestamp(),
             shards: Vec::new(),
-            capacity_bytes: 10 * 1024 * 1024 * 1024, // 10GB default
+            capacity_bytes: (10u64 * 1024 * 1024 * 1024) as usize, // 10GB default
             used_bytes: 0,
         };
